fix(ci): python-full-coverage — align tests with attempt-start HEAD probe and file limits
Split comment_verdict residue helpers and verdict-retry part_002 tests to satisfy
the 1500-line maintainability gate, and update FakeCommandRunner queues for the new
attempt-start rev-parse and live-head rollback paths in verdict/commit-sink tests.
